WARNING
This program binds a full SSH server to port 22. It has to be run as
route, it does absolutely no authentication, and it gives whoever
connects to it a root terminal. It is here for proof of concepts while
the client side is being developed in parallel. DO NOT RUN IT WHILE
EXPOSED TO THE INTERNET.

After 10 minutes with a public IP address, you will be scanned by at
least of a handful of botnets, at least a couple of hackers trying
invalid users, and if you have this attached, THEY WILL ALL GET IN.