Fix looping in general for client and server NOTE:
Right now it seems to rely on SSH to continually call read
SSH seems to sit on READ  no matter what
Doesn't seem to send anything down the line
But WSConn will keep on dealing with the text as long as SSH sits on
READ
If I call ssh.close() I think it just tries to close underlying
connection (which it can't) so it just continues to sit on read?
If I could pull it off, I would, but something would need to sit on
read.
Also ping.write() seems to take a minute to realize that it's failed
wsconn realizes the pipe is broken 2 or 3 pings early.
All stuff to consider.